refactor: ViewModel로 분리함에 따라, TimelineUseCase 제거
